refactor: upgrade nwp500-python to 7.3.3 and implement unit-aware fixes
- Upgrade nwp500-python dependency to 7.3.3
- Update water_heater.py to pass temperature as float to set_temperature
- Update __init__.py to use new unit-agnostic build_reservation_entry API
- Pass device-specific temperature limits (min/max) to reservation builder
- Update tests to reflect API changes

+++ custom_components/nwp500/water_heater.py
@@ -330,7 +330,7 @@ class NWP500WaterHeater(NWP500Entity, WaterHeaterEntity):  # type: ignore[report
             return
 
         success = await self.coordinator.async_control_device(
-            self.mac_address, "set_temperature", temperature=int(temperature)
+            self.mac_address, "set_temperature", temperature=float(temperature)
         )
 
         if success:
